Conflicts:
	src/minecraft/micdoodle8/mods/galacticraft/mars/wgen/GCMarsChunkProvider.java
	src/minecraft/micdoodle8/mods/galacticraft/moon/wgen/GCMoonChunkProvider.java

+++ src/minecraft/micdoodle8/mods/galacticraft/europa/dimension/GCEuropaWorldProvider.java
@@ -10,11 +10,11 @@ import net.minecraft.util.Vec3;
 import net.minecraft.world.WorldProvider;
 import net.minecraft.world.chunk.Chunk;
 import net.minecraft.world.chunk.IChunkProvider;
-import cpw.mods.fml.relauncher.SideOnly;
 import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 
 /**
- * Copyright 2012, micdoodle8
+ * Copyright 2012-2013, micdoodle8
  * 
  *  All rights reserved.
  *
@@ -70,7 +70,7 @@ public class GCEuropaWorldProvider extends WorldProvider implements IGalacticraf
 	@Override
     public Vec3 getSkyColor(Entity cameraEntity, float partialTicks)
     {
-        return this.worldObj.getWorldVec3Pool().getVecFromPool((double)0, (double)0, (double)0);
+        return this.worldObj.getWorldVec3Pool().getVecFromPool(0, 0, 0);
     }
 	
 	@Override
@@ -197,7 +197,7 @@ public class GCEuropaWorldProvider extends WorldProvider implements IGalacticraf
     }
 
 	@Override
-	public String getDimensionName() 
+	public String getDimensionName()
 	{
 		return "Europa";
 	}
@@ -227,13 +227,13 @@ public class GCEuropaWorldProvider extends WorldProvider implements IGalacticraf
     }
 
 	@Override
-	public float getGravity() 
+	public float getGravity()
 	{
 		return 0.069F;
 	}
 
 	@Override
-	public float getMeteorFrequency() 
+	public float getMeteorFrequency()
 	{
 		return 0F;
 	}
