+++ src/minecraft/micdoodle8/mods/galacticraft/mars/dimension/GCMarsWorldProvider.java
@@ -11,11 +11,11 @@ import net.minecraft.util.Vec3;
 import net.minecraft.world.WorldProvider;
 import net.minecraft.world.chunk.Chunk;
 import net.minecraft.world.chunk.IChunkProvider;
-import cpw.mods.fml.relauncher.SideOnly;
 import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 
 /**
- * Copyright 2012, micdoodle8
+ * Copyright 2012-2013, micdoodle8
  * 
  *  All rights reserved.
  *
@@ -71,7 +71,7 @@ public class GCMarsWorldProvider extends WorldProvider implements IGalacticraftW
 	@Override
     public Vec3 getSkyColor(Entity cameraEntity, float partialTicks)
     {
-        return this.worldObj.getWorldVec3Pool().getVecFromPool((double)0, (double)0, (double)0);
+        return this.worldObj.getWorldVec3Pool().getVecFromPool(0, 0, 0);
     }
     
     @Override
@@ -197,7 +197,7 @@ public class GCMarsWorldProvider extends WorldProvider implements IGalacticraftW
     }
 
 	@Override
-	public String getDimensionName() 
+	public String getDimensionName()
 	{
 		return "Mars";
 	}
@@ -227,13 +227,13 @@ public class GCMarsWorldProvider extends WorldProvider implements IGalacticraftW
     }
 
 	@Override
-	public float getGravity() 
+	public float getGravity()
 	{
 		return 0.005F;
 	}
 
 	@Override
-	public float getMeteorFrequency() 
+	public float getMeteorFrequency()
 	{
 		return 0F;
 	}
